+++ d20app/__init__.py
@@ -6,4 +6,4 @@ the cue that it's OK to give the cat a treat. Configured entirely from a local
 web GUI.
 """
 
-__version__ = "0.3.2"
+__version__ = "0.3.3"